requirements.txt
python-telegram-bot==20.7
openai-whisper==20231117
pydub==0.25.1
torch==2.1.0
ffmpeg-python==0.2.0